test(server): add updateUserSummary tests
