+++ apps/server/src/infrastructure/transport/graphql/resolvers/user.resolvers.test.js
@@ -1,27 +1,9 @@
+const { createUserHelper, updateUserSummaryHelper } = require('../../../../utils/tests/user.resolver.helpers');
 const testServer = require('../server')
 
 describe('User Resolvers', () => {
-  it('returns hello with the provided name', async () => {
-    const result = await testServer.executeOperation({
-      query: `
-        mutation CreateUser($user: UserInput) {
-          createUser(user: $user) {
-            id
-            email
-            username
-            photo
-            summary
-          }
-        }
-      `,
-      variables: {
-        user: {
-          email: 'any_email@mail.com',
-          username: 'any_username',
-          password: 'any_password'
-        }
-      },
-    });
+  it('should return a new user when createUser is called', async () => {
+    const result = await testServer.executeOperation(createUserHelper())
     expect(result.errors).toBeUndefined()
     expect(result.data?.createUser).toEqual({
       id: '1',
@@ -31,22 +13,11 @@ describe('User Resolvers', () => {
       summary: null
     })
   })
+  
+  it('should return an user with updated summary when UpdateUserSummary is called', async () => {
+    
+    const { data: { createUser} } = await testServer.executeOperation(createUserHelper())
+    const { data: { updateUserSummary} } = await testServer.executeOperation(updateUserSummaryHelper({ updateUserSummaryId: createUser.id, summary: 'any_summary' }))
+    expect(updateUserSummary.summary).toBe('any_summary')
+  })
 })
-
-
-/*
-mutation CreateUser($user: UserInput) {
-  createUser(user: $user) {
-    id
-    email
-    username
-    photo
-    summary
-  }
-}
-
-"user": {
-    "email": "demo1@demo.com",
-    "password": "1234",
-    "usern
-*/
